requirements.txt
youtube_transcript_api
streamlit
google-generativeai
pathlib
langchain
PyPDF2
faiss-cpu
langchain_google_genai